### 2016-05-08 #43 @masalmon
geoparser: geoparsing with the geoparser.io API
* 1. What does this package do? (explain in 50 words or less)
This package is an interface to the [geoparser.io API](https://geoparser.io) that identifies places mentioned in text, disambiguates those places, and returns data about the places found in the text.
* 2. Paste the full DESCRIPTION file inside a code block below.

```
Package: geoparser
Title: Interface to the Geoparser.io API
Version: 0.1.0
Authors@R: person("Maëlle", "Salmon", email = "maelle.salmon@yahoo.se", role = c("aut", "cre"))
Description: This packages accesses the Geoparser.io API which is a web service
    that identifies places mentioned in text, disambiguates those places, and
    returns detailed data about the places found in the text.
Suggests: testthat,
Depends: dplyr, httr, jsonlite, lazyeval, tidyr

```

* 3. URL for the package (the development repository, not a stylized html page)
https://github.com/masalmon/geoparser
* 4. What data source(s) does it work with (if applicable)?
The geoparser.io API which is version beta. They said "We expect it to remain largely the same for the immediate future. We'd try to make any changes backwards compatible, and support older versions of the API for an extended period of time to allow users to migrate at their convenience. However, the most likely things to change are:
 1. The possible values for Feature.properties.type in the GeoJSON response data
 2.   How we calculate values for Feature.properties.confidence in the GeoJSON response data
 3.   The endpoint URI for the API itself"
* 5. Who is the target audience?
People interested in Natural Language Processing, or anyone with text where they would like to identify geographical information.
* 6. Are there other R packages that accomplish the same thing? If so, what is different about yours?
I don't think so. Named Entity Recognition (including placenames) in R is currently offered by the openNLP package that requires Java but then the results are not geolocated.
* 7. Check the box next to each policy below, confirming that you agree. These are mandatory.
    * [x ] This package does not violate the Terms of Service of any service it interacts with.
    * [ x] The repository has continuous integration with Travis CI and/or another service
    * [ x] The package contains a vignette
    * [ x] The package contains a reasonably complete README with `devtools` install instructions
    * [ x] The package contains unit tests
    * [ x] The package only exports functions to the NAMESPACE that are intended for end users
* 8. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.
    * [ ] Are there any package dependencies not on CRAN?
    * [ x] Do you intend for this package to go on CRAN?
    * [ x] Does the package have a CRAN accepted license?
    * [ ] Did `devtools::check()` produce any errors or warnings? If so paste them below.
* 9. Please add explanations below for any exceptions to the above:
* 10. If this is a resubmission following rejection, please explain the change in circumstances.
