### 2016-05-09 #43 @noamross
Reviewers: @hrbrmstr
Due date: 2016-05-30
