### 2016-05-30 #43 @sckott
@hrbrmstr
Due date: 2016-05-30 - hey there, it's been 21 days, please get your review in soon, thanks :smiley_cat: (ropensci-bot)
