### 2016-06-04 #43 @sckott
@hrbrmstr
Due date: 2016-05-30 - hey there, it's been 26 days, please get your review in soon, thanks :smiley_cat: (ropensci-bot)
