### 2016-06-10 #43 @masalmon
Thank you so much, @hrbrmstr! I look forward to making all the suggested changes! :smile_cat:
