### 2016-06-10 #43 @noamross
Thanks for the thorough review @hrbrmstr.
