### 2016-06-10 #43 @masalmon
I don't think you owed me the PR, @hrbrmstr so many thanks! It really was useful! I have learnt a lot. :+1:

Reg.: "Nice use of the "standard"/modern R way to access APIs requring an API key." At least part of it is my copy-pasting the function @noamross had added to `opencage`. :smile:

* The encoding thing was in the sample code the API contact person had sent to me when advertising for their service (I had commented on a GH issue about geoparsing and that's how they found me)

I'm going to summarize what I did:

* I have looked at all comments and changes. Thank you!

* I have added you as `rev` and `ctb` in the authors list.

* I have commented my code a bit more. I am already future me from the perspective of the coder who submitted the package a month ago... This coder was a bit too optimistic about her memory.

* I have replaced some apply-things by purrr functions but honestly I will need to come back to it later because it's clumsy and when I use map(blabla, unlist), I don't make any progress compared to the earlier version.

* The function is now vectorized! For identifying each input text in the output I added a column with the MD5 hash of the text. Maybe there is a better solution? I added the MD5 hash because digest seemed to be a light dependency and I did not want to add the whole text.

* I added the Example request number 2 from the site in the vignette and in the README, thanks!

* "8,000 vs 8,192" change made.

* I have used mutate_ instead of "$" for latitude/longitude. Other $ are for lists.

* I now use `purrr::safely`... Is what I wrote what you had in mind in your comment?

Once you all (@hrbrmstr,  @noamross & @sckott  ) will be happy with the package, I'll transfer the repo to ropenscilabs but also submit it to CRAN. I had discussed the possible API changes with geoparser.io and I think it's ok to submit this version. I have added the API version number to DESCRIPTION.
