### 2016-06-16 #43 @noamross
Thanks, @masalmon! @hrbrmstr, please let us know if the response addresses everything to your satisfaction.
