### 2016-06-16 #43 @hrbrmstr
aye. i also re-looked at the code and it's #spiffy.

On Thu, Jun 16, 2016 at 11:10 AM, Noam Ross <notifications@github.com>
wrote:

> Thanks, @masalmon <https://github.com/masalmon>! @hrbrmstr
> <https://github.com/hrbrmstr>, please let us know if the response
> addresses everything to your satisfaction.
>
