### 2016-06-16 #43 @noamross
Thank you @hrbrmstr.

I have done an editors' check and you are approved!  Please go ahead and transfer the repository and add the rOpenSci footer. 👏
