### 2016-06-16 #43 @masalmon
Transfer done. Next step will be the CRAN submission!
